minor UI tweaks and chat implementation in chatviewmodel

+++ app/src/main/java/com/frangerapp/franger/ui/home/OutgoingListItemUiState.java
@@ -14,7 +14,7 @@ public class OutgoingListItemUiState {
     public String userId;
     public String imageUrl;
     public User user;
-    public int unreadCount = 0;
+    public long unreadCount = 0;
     public Date timeStamp;
     public OutgoingGroupItemClickHandler handler;
     public long messageId;
@@ -48,11 +48,11 @@ public class OutgoingListItemUiState {
         this.user = user;
     }
 
-    public int getUnreadCount() {
+    public long getUnreadCount() {
         return unreadCount;
     }
 
-    public void setUnreadCount(int unreadCount) {
+    public void setUnreadCount(long unreadCount) {
         this.unreadCount = unreadCount;
     }
